requirements.txt
Flask==2.3.3
python-slugify==8.0.4
Werkzeug==2.3.7